integration: rename load_balancing.rs -> latency_awareness.rs
The only test that is left there regards latency latency_awareness,
so such rename makes sense. And module_inception is no longer there.
